:sparkles: #4 pixiJS 배경에 세팅
